(refactor) Supplier validation and proyect structure changes
