+++ src/main/java/com/utn/ProgIII/dto/UpdateProductSupplierDTO.java
@@ -1,0 +1,10 @@
+package com.utn.ProgIII.dto;
+
+import java.math.BigDecimal;
+
+public record UpdateProductSupplierDTO(
+
+        BigDecimal cost,
+
+        BigDecimal profitMargin) {
+}
