+++ src/main/java/com/utn/ProgIII/repository/ProductSupplierRepository.java
@@ -1,7 +1,7 @@
 package com.utn.ProgIII.repository;
 
 import com.utn.ProgIII.model.Product.Product;
-import com.utn.ProgIII.model.ProductSupplier.dto.ExtendedProductDTO;
+import com.utn.ProgIII.dto.ExtendedProductDTO;
 import com.utn.ProgIII.model.ProductSupplier.ProductSupplier;
 import com.utn.ProgIII.model.Supplier.Supplier;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -12,7 +12,7 @@ import java.util.List;
 
 public interface ProductSupplierRepository extends JpaRepository<ProductSupplier, Long> {
 
-    @Query("SELECT NEW com.utn.ProgIII.model.ProductSupplier.dto.ExtendedProductDTO(" +
+    @Query("SELECT NEW com.utn.ProgIII.dto.ExtendedProductDTO(" +
             "p.idProduct, p.name, ps.cost, ps.profitMargin, ps.price, p.status) " +
             "FROM ProductSupplier ps JOIN ps.product p " +
             "WHERE ps.supplier.id = :idSupplier")
