+++ src/main/java/com/utn/ProgIII/service/implementations/ProductSupplierServiceImpl.java
@@ -1,12 +1,12 @@
 package com.utn.ProgIII.service.implementations;
 
+import com.utn.ProgIII.dto.*;
 import com.utn.ProgIII.exceptions.ProductNotFoundException;
 import com.utn.ProgIII.exceptions.ProductSupplierNotExistException;
 import com.utn.ProgIII.exceptions.SupplierNotFoundException;
 import com.utn.ProgIII.mapper.ProductSupplierMapper;
 import com.utn.ProgIII.model.Product.Product;
 import com.utn.ProgIII.model.ProductSupplier.*;
-import com.utn.ProgIII.model.ProductSupplier.dto.*;
 import com.utn.ProgIII.model.Supplier.Supplier;
 import com.utn.ProgIII.repository.ProductRepository;
 import com.utn.ProgIII.repository.ProductSupplierRepository;
